requirements.txt
pillow==9.1.1
requests==2.27.1
telethon==1.24.0